without libft

+++ srcs/ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: mvpee <mvpee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/11 09:32:00 by mvpee             #+#    #+#             */
-/*   Updated: 2023/10/15 17:02:20 by mvpee            ###   ########.fr       */
+/*   Updated: 2023/11/03 19:25:12 by mvpee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
 	size_t	i;
-	int		len;
+	size_t	len;
 
 	i = -1;
 	len = ft_strlen(src);
